web/requirements.txt
django
djangorestframework
celery
django-celery-results
gunicorn
redis